feat : 삭제 시 삭제 경고 문구 내보내기 #15

+++ src/client/javascripts/main/detailDiary.js
@@ -11,8 +11,12 @@ class DetailDiary {
   deleteThisDiary() {
     return function () {
       const diaryId = $('diaryId').value;
-      deleteData(`/api/diaries/${diaryId}`);
-      // window.location.reload();
+      const returned = confirm('정말로 삭제하시겠습니까?');
+
+      if (returned) {
+        deleteData(`/api/diaries/${diaryId}`);
+        window.location.reload();
+      }
     };
   }
